requirements.txt
django==3.2
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.3
pillow==10.0.0
django-mail-templated==2.6.5
six==1.16.0
django-render-partial==0.4
django-star-ratings==0.9.2